backend/requirements.txt
flask
langchain
langchain_ollama
mysql-connector-python
pypdf
